parallelisation

+++ ams.py
@@ -4,7 +4,7 @@ import os
 import shutil
 import json
 from ase.io import Trajectory, read, write
-from ase.parallel import parprint, paropen, world
+from ase.parallel import parprint, paropen, world, barrier
 
 
 class CollectiveVariables:
@@ -711,14 +711,16 @@ class InitialConditionsSampler:
 class AMS:
     """Running AMS"""
 
-    def __init__(self,
-                 n_rep: int,
-                 k_min: int,
-                 dyn,
-                 xi,
-                 cv_interval: int = 1,
-                 rc_threshold: float = 10 ** (-3),
-                 save_all: bool = False):
+    def __init__(
+        self,
+        n_rep,
+        k_min,
+        dyn,
+        xi,
+        cv_interval=1,
+        rc_threshold=1.0e-3,
+        save_all=False,
+    ):
         """
         Parameters:
 
@@ -829,9 +831,9 @@ class AMS:
             z = self.xi.rc(self.dyn.atoms)
             f = paropen(self.alive_traj_dir + "/rc_rep_" + str(i) + ".txt", "a")
             if self.xi.in_r(self.dyn.atoms):
-                z = -1.0 * 10**8
+                z = -1.0e8
             elif self.xi.in_p(self.dyn.atoms):
-                z = 1.0 * 10**8
+                z = 1.0e8
             f.write(str(z) + "\n")
             if z >= self.z_maxs[i]:
                 self.z_maxs[i] = z
@@ -845,7 +847,7 @@ class AMS:
             raise ValueError("""The directory of alive trajectories is not defined ! Call ams.set_alive_traj_dir""")
         if self.save_all is True and self.non_reac_traj_dir is None:
             raise ValueError("""The directory of non reactive trajectories is not defined ! Call ams.set_non_reac_traj_dir""")
-        self.z_maxs = (np.ones(self.n_rep) * (-1 * 10**8)).tolist()
+        self.z_maxs = (np.ones(self.n_rep) * (-1.0e8)).tolist()
         i = 0
         if len(os.listdir(self.alive_traj_dir)) == 0:
             rep = -1
@@ -856,16 +858,19 @@ class AMS:
                 self.z_maxs[j] = np.max(np.loadtxt(self.alive_traj_dir + "/rc_rep_" + str(i) + ".txt"))
         while i < self.n_rep:
             if i > rep:
-                ini_cond = np.random.choice(os.listdir(self.ini_cond_dir))
+                ini_cond = np.random.choice([ini for ini in os.listdir(self.ini_cond_dir) if "_used" not in ini])
                 atoms = read(self.ini_cond_dir + "/" + ini_cond, index=0)
-                os.remove(self.ini_cond_dir + "/" + ini_cond)
+                if world.rank == 0:
+                    filename, file_extension = os.path.splitext(ini_cond)
+                    os.rename(self.ini_cond_dir + "/" + ini_cond, self.ini_cond_dir + "/" + filename + "_used" + file_extension)
                 self.dyn.atoms.set_scaled_positions(atoms.get_scaled_positions())
                 self.dyn.atoms.set_momenta(atoms.get_momenta())
                 self.dyn.atoms.set_calculator(self.calc)
                 traj = self.dyn.closelater(Trajectory(filename=self.alive_traj_dir + "/rep_" + str(i) + ".traj", mode="w", atoms=self.dyn.atoms))
             else:
                 read_traj = read(filename=self.alive_traj_dir + "/rep_" + str(i) + ".traj", format="traj", index=":")
-                os.remove(self.alive_traj_dir + "/rep_" + str(i) + ".traj")
+                if world.rank == 0:
+                    os.remove(self.alive_traj_dir + "/rep_" + str(i) + ".traj")
                 self.dyn.atoms.set_scaled_positions(read_traj[-1].get_scaled_positions())
                 self.dyn.atoms.set_momenta(read_traj[-1].get_momenta())
                 self.dyn.atoms.set_calculator(self.calc)
@@ -875,9 +880,9 @@ class AMS:
             z = self.xi.rc(self.dyn.atoms)
             f = paropen(self.alive_traj_dir + "/rc_rep_" + str(i) + ".txt", "a")
             if self.xi.in_r(self.dyn.atoms):
-                z = -1.0 * 10**8
+                z = -1.0e8
             elif self.xi.in_p(self.dyn.atoms):
-                z = 1.0 * 10**8
+                z = 1.0e8
             f.write(str(z) + "\n")
             f.close()
             self.dyn.attach(traj.write, interval=self.cv_interval)
@@ -894,15 +899,16 @@ class AMS:
 
     def _branch_replica(self, i, j, z_kill):
         """Branch replica i by copying replica j until z_kill and run the dynamics until it reaches either R or P"""
-        if self.save_all:
+        if self.save_all and world.rank == 0:
             os.system("cp " + self.alive_traj_dir + "/rep_" + str(i) + ".traj " + self.non_reac_traj_dir + "/rep_" + str(i) + "_killed_at_" + str(self.ams_it) + ".traj")
             json_file = paropen(self.non_reac_traj_dir + "/rep_" + str(i) + "_killed_at_" + str(self.ams_it) + "_weights.txt", "w")
             weights = {"weights": self.rep_weights[i]}
             json.dump(weights, json_file, indent=4)
             json_file.close()
         self.rep_weights[i] = []
-        os.remove(self.alive_traj_dir + "/rep_" + str(i) + ".traj")
-        os.remove(self.alive_traj_dir + "/rc_rep_" + str(i) + ".txt")
+        if world.rank == 0:
+            os.remove(self.alive_traj_dir + "/rep_" + str(i) + ".traj")
+            os.remove(self.alive_traj_dir + "/rc_rep_" + str(i) + ".txt")
         read_traj = read(filename=self.alive_traj_dir + "/rep_" + str(j) + ".traj", format="traj", index=":")
         k = 0
         traj = self.dyn.closelater(Trajectory(filename=self.alive_traj_dir + "/rep_" + str(i) + ".traj", mode="w", atoms=read_traj[k]))
@@ -925,7 +931,7 @@ class AMS:
 
     def _iteration(self):
         """Perform one iteration of the AMS algorithm"""
-        if np.min(self.z_maxs) == 1.0 * 10**8:
+        if np.min(self.z_maxs) >= 1.0e8:
             self.finished = True
             self.success = True
             self._write_checkpoint()
@@ -962,11 +968,11 @@ class AMS:
             rc_traj = np.loadtxt(self.alive_traj_dir + "/rc_rep_" + str(i) + ".txt")
             if len(rc_traj.shape) == 0:
                 rc_traj.reshape([1])
-            if np.max(rc_traj) <= z_kill and (rc_traj[-1] == -1.0 * 10**8 or rc_traj[-1] == 1.0 * 10**8):
+            if np.max(rc_traj) <= z_kill and (rc_traj[-1] <= -1.0e8 or rc_traj[-1] >= 1.0e8):
                 j = np.random.choice(alive)
                 self._branch_replica(i, j, z_kill)
                 self._until_r_or_p(i)
-            elif np.max(rc_traj) > z_kill and not (rc_traj[-1] == -1.0 * 10**8 or rc_traj[-1] == 1.0 * 10**8):
+            elif np.max(rc_traj) > z_kill and not (rc_traj[-1] <= -1.0e8 or rc_traj[-1] >= 1.0e8):
                 read_traj = read(filename=self.alive_traj_dir + "/rep_" + str(i) + ".traj", format="traj", index=":")
                 self.dyn.atoms.set_scaled_positions(read_traj[-1].get_scaled_positions())
                 self.dyn.atoms.set_momenta(read_traj[-1].get_momenta())
@@ -1006,6 +1012,7 @@ class AMS:
         if os.path.exists(self.progress_dir + "/ams_checkpoint.txt"):
             parprint("Read checkpoint")
             self._read_checkpoint()
+        barrier()  # Wait for all threads to read checkpoint
         if not self.initialized:
             self.dyn.observers = []
             self._initialize()
